proyecto_final_DAJ/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
greenlet==3.0.3
idna==3.7
numpy==1.26.4
pandas==2.2.2
pandas-stubs==2.2.2.240514
psycopg2==2.9.9
pyarrow==16.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
SQLAlchemy==2.0.30
tqdm==4.66.4
types-pytz==2024.1.0.20240417
types-requests==2.32.0.20240523
types-tqdm==4.66.0.20240417
typing_extensions==4.12.0
tzdata==2024.1
urllib3==2.2.1
